requirements.txt
pika>=1.3.0
python-dotenv>=0.19.0
